requirements.txt
streamlit==1.27.0
python-dotenv==1.0.1
chromadb==0.4.3
sentence-transformers==2.2.2
langchain==1.0.5
PyPDF2==2.10.0
python-docx==0.8.12
pandas==2.1.1
matplotlib==3.8.0
duckdb==0.8.1

